backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# LangChain 1.x
langchain==0.1.4
langchain-core==0.1.18
langchain-anthropic==0.1.1

# LLM Providers
openai==1.10.0

# File processing
python-docx==1.1.0
PyPDF2==3.0.1
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
cryptography==41.0.7

# Async
aiofiles==23.2.1
